Refactor markdown components structure
