+++ layouts/ApiLayout.re
@@ -13,6 +13,7 @@ open Text;
 module Link = Next.Link;
 
 module MainMd = {
+  open Markdown;
   module P = {
     [@react.component]
     let make = (~children) => {
@@ -32,18 +33,18 @@ module MainMd = {
   let components =
     Mdx.Components.t(
       ~p=P.make,
-      ~li=Md.Li.make,
+      ~li=Li.make,
       ~h1=H1.make,
       ~h2=H2.make,
       ~h3=H3.make,
       ~h4=H4.make,
       ~h5=H5.make,
-      ~ul=Md.Ul.make,
-      ~ol=Md.Ol.make,
-      ~a=Md.A.make,
-      ~pre=Md.Pre.make,
-      ~inlineCode=Md.InlineCode.make,
-      ~code=Md.Code.make,
+      ~ul=Ul.make,
+      ~ol=Ol.make,
+      ~a=A.make,
+      ~pre=Pre.make,
+      ~inlineCode=InlineCode.make,
+      ~code=Code.make,
       (),
     );
 };
