+++ layouts/SidebarLayout.re
@@ -16,190 +16,8 @@ hljs.registerLanguage('reason', reasonHighlightJs);
 |};
 
 open Util.ReactStuff;
-open Text;
 module Link = Next.Link;
 
-/*
-    We use some custom markdown styling for the Belt docs to make
-    it easier on the eyes
- */
-module ApiMd = {
-  module InvisibleAnchor = {
-    [@react.component]
-    let make = (~id: string) => {
-      let style =
-        ReactDOMRe.Style.make(~position="absolute", ~top="-1rem", ());
-      <span className="relative" ariaHidden=true> <a id style /> </span>;
-    };
-  };
-
-  module H1 = {
-    [@react.component]
-    let make = (~children) => {
-      <h1
-        className="text-6xl leading-1 mb-2 font-sans font-medium text-night-dark">
-        children
-      </h1>;
-    };
-  };
-
-  module H2 = {
-    // We will currently hide the headline, to keep the structure,
-    // but having an Elm like documentation
-    [@react.component]
-    let make = (~children) => {
-      <>
-        // Here we know that children is always a string (## headline)
-        <InvisibleAnchor id={children->Unsafe.elementAsString} />
-        <div className="border-b border-gray-200 my-20" />
-      </>;
-    };
-  };
-
-  module Pre = {
-    [@react.component]
-    let make = (~children) => {
-      <pre className="mt-2 mb-4 block"> children </pre>;
-    };
-  };
-
-  module P = {
-    [@react.component]
-    let make = (~children) => {
-      <p className="mt-3 leading-4 text-night"> children </p>;
-    };
-  };
-
-  let components =
-    Mdx.Components.t(
-      ~p=P.make,
-      ~li=Md.Li.make,
-      ~h1=H1.make,
-      ~h2=H2.make,
-      ~h3=H3.make,
-      ~h4=H4.make,
-      ~h5=H5.make,
-      ~ul=Md.Ul.make,
-      ~ol=Md.Ol.make,
-      ~a=Md.A.make,
-      ~thead=Md.Thead.make,
-      ~th=Md.Th.make,
-      ~td=Md.Td.make,
-      ~pre=Pre.make,
-      ~inlineCode=Md.InlineCode.make,
-      ~code=Md.Code.make,
-      (),
-    );
-};
-
-module ProseMd = {
-  module Anchor = {
-    [@react.component]
-    let make = (~id: string) => {
-      let style =
-        ReactDOMRe.Style.make(~position="absolute", ~top="-7rem", ());
-      <span className="inline group relative">
-        <a
-          className="invisible text-night-light opacity-50 text-inherit hover:opacity-100 hover:text-night-light hover:cursor-pointer group-hover:visible"
-          href={"#" ++ id}>
-          {j|#|j}->s
-        </a>
-        <a style id />
-      </span>;
-    };
-  };
-
-  module H2 = {
-    [@react.component]
-    let make = (~children) => {
-      <>
-        // Here we know that children is always a string (## headline)
-        <h2
-          className="group mt-12 text-3xl leading-1 font-sans font-medium text-night-dark">
-          <span className="-ml-8 pr-2">
-            <Anchor id={children->Unsafe.elementAsString} />
-          </span>
-          children
-        </h2>
-      </>;
-    };
-  };
-
-  module H3 = {
-    [@react.component]
-    let make = (~children) => {
-      <h3
-        className="group text-xl mt-12 leading-3 font-sans font-medium text-night-darker">
-        <span className="-ml-6 pr-2">
-          <Anchor id={children->Unsafe.elementAsString} />
-        </span>
-        children
-      </h3>;
-    };
-  };
-
-  module H4 = {
-    [@react.component]
-    let make = (~children) => {
-      <h4
-        className="group text-lg mt-12 leading-2 font-sans font-semibold text-night-dark">
-        <span className="-ml-5 pr-2">
-          <Anchor id={children->Unsafe.elementAsString} />
-        </span>
-        children
-      </h4>;
-    };
-  };
-
-  module H5 = {
-    [@react.component]
-    let make = (~children) => {
-      <h5
-        className="group mt-12 text-xs leading-2 font-sans font-semibold uppercase tracking-wide">
-        <span className="-ml-5 pr-2">
-          <Anchor id={children->Unsafe.elementAsString} />
-        </span>
-        children
-      </h5>;
-    };
-  };
-
-  module Pre = {
-    [@react.component]
-    let make = (~children) => {
-      <pre className="mt-2 mb-4 block"> children </pre>;
-    };
-  };
-
-  module P = {
-    [@react.component]
-    let make = (~children) => {
-      <p className="text-inherit mt-3 leading-4 text-night"> children </p>;
-    };
-  };
-
-  let components =
-    Mdx.Components.t(
-      ~p=P.make,
-      ~li=Md.Li.make,
-      ~h1=ApiMd.H1.make,
-      ~h2=H2.make,
-      ~h3=H3.make,
-      ~h4=H4.make,
-      ~h5=H5.make,
-      ~ul=Md.Ul.make,
-      ~ol=Md.Ol.make,
-      ~a=Md.A.make,
-      ~table=Md.Table.make,
-      ~thead=Md.Thead.make,
-      ~th=Md.Th.make,
-      ~td=Md.Td.make,
-      ~pre=Pre.make,
-      ~inlineCode=Md.InlineCode.make,
-      ~code=Md.Code.make,
-      (),
-    );
-};
 
 module UrlPath = {
   /*
@@ -587,7 +405,7 @@ module MobileDrawerButton = {
 let make =
     (
       ~theme: ColorTheme.t,
-      ~components=ApiMd.components,
+      ~components: Mdx.Components.t,
       // (Sidebar, toggleSidebar)
       ~sidebar: (React.element, unit => unit),
       ~breadcrumbs: option(list(UrlPath.breadcrumb))=?,
