+++ styles/_markdown.css
@@ -10,7 +10,7 @@
   @apply text-night-dark font-semibold inline-block w-6 -ml-6;
 }
 
-.md-ul > .md-li p::before, .md-ul > .md-li > a::before {
+.md-ul > .md-li p::before {
   content: "•";
   @apply inline-block w-4 -ml-4 pr-2;
 }
